src/video_indexer/requirements.txt
fastapi==0.112.0
python-dotenv==1.0.1
pytubefix==6.9.2